backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==1.1.0
openai==1.99.1
pytest==7.4.3
httpx==0.25.2 